Remove server-specific error handling from common, reduce public plugin bundle
