+++ x-pack/plugins/cases/server/client/attachments/add.ts
@@ -10,7 +10,7 @@ import { SavedObjectsUtils } from '@kbn/core/server';
 import { validateMaxUserActions } from '../../../common/utils';
 import { AttachmentRequestRt } from '../../../common/types/api';
 import type { Case } from '../../../common/types/domain';
-import { decodeWithExcessOrThrow } from '../../../common/api';
+import { decodeWithExcessOrThrow } from '../../common/runtime_types';
 import { CaseCommentModel } from '../../common/models';
 import { createCaseError } from '../../common/error';
 import type { CasesClientArgs } from '..';
