+++ x-pack/plugins/cases/server/client/cases/delete.ts
@@ -12,7 +12,7 @@ import type { SavedObjectsBulkDeleteObject } from '@kbn/core/server';
 import type { FileServiceStart } from '@kbn/files-plugin/server';
 import type { CasesDeleteRequest } from '../../../common/types/api';
 import { CasesDeleteRequestRt } from '../../../common/types/api';
-import { decodeWithExcessOrThrow } from '../../../common/api';
+import { decodeWithExcessOrThrow } from '../../common/runtime_types';
 import {
   CASE_COMMENT_SAVED_OBJECT,
   CASE_SAVED_OBJECT,
