+++ x-pack/plugins/cases/server/client/cases/get.ts
@@ -26,7 +26,7 @@ import {
   GetReportersResponseRt,
   GetTagsResponseRt,
 } from '../../../common/types/api';
-import { decodeWithExcessOrThrow } from '../../../common/api';
+import { decodeWithExcessOrThrow, decodeOrThrow } from '../../common/runtime_types';
 import { createCaseError } from '../../common/error';
 import { countAlertsForID, flattenCaseSavedObject, countUserAttachments } from '../../common/utils';
 import type { CasesClientArgs } from '..';
@@ -34,7 +34,6 @@ import { Operations } from '../../authorization';
 import { combineAuthorizedAndOwnerFilter } from '../utils';
 import { CasesService } from '../../services';
 import type { CaseSavedObjectTransformed } from '../../common/types/case';
-import { decodeOrThrow } from '../../../common/api/runtime_types';
 import { CaseRt } from '../../../common/types/domain';
 
 /**
