+++ x-pack/plugins/cases/server/client/cases/replace_custom_field.ts
@@ -6,7 +6,6 @@
  */
 
 import Boom from '@hapi/boom';
-
 import type { CasesClient, CasesClientArgs } from '..';
 
 import type { CustomFieldPutRequest } from '../../../common/types/api';
@@ -14,10 +13,9 @@ import { CustomFieldPutRequestRt, CaseRequestCustomFieldsRt } from '../../../com
 import { Operations } from '../../authorization';
 import { createCaseError } from '../../common/error';
 import { validateMaxUserActions } from '../../../common/utils/validators';
-import { decodeOrThrow } from '../../../common/api/runtime_types';
+import { decodeWithExcessOrThrow, decodeOrThrow } from '../../common/runtime_types';
 import type { CaseCustomField } from '../../../common/types/domain';
 import { CaseCustomFieldRt } from '../../../common/types/domain';
-import { decodeWithExcessOrThrow } from '../../../common/api';
 import { validateCustomFieldTypesInRequest } from './validators';
 import type { UserActionEvent } from '../../services/user_actions/types';
