+++ x-pack/plugins/cases/server/client/configure/update_mappings.ts
@@ -8,7 +8,7 @@
 import { ACTION_SAVED_OBJECT_TYPE } from '@kbn/actions-plugin/server';
 import type { ConnectorMappingResponse } from '../../../common/types/api';
 import { ConnectorMappingResponseRt } from '../../../common/types/api';
-import { decodeOrThrow } from '../../../common/api';
+import { decodeOrThrow } from '../../common/runtime_types';
 import { createCaseError } from '../../common/error';
 import type { CasesClientArgs } from '..';
 import type { UpdateMappingsArgs } from './types';
