+++ x-pack/plugins/cases/server/client/user_actions/stats.ts
@@ -7,7 +7,7 @@
 
 import type { CaseUserActionStatsResponse } from '../../../common/types/api';
 import { CaseUserActionStatsResponseRt } from '../../../common/types/api';
-import { decodeOrThrow } from '../../../common/api';
+import { decodeOrThrow } from '../../common/runtime_types';
 import { createCaseError } from '../../common/error';
 import type { CasesClientArgs } from '..';
 import type { UserActionGet } from './types';
