+++ x-pack/plugins/cases/server/common/types/connector_mappings.test.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import { decodeOrThrow } from '../../../common/api/runtime_types';
+import { decodeOrThrow } from '../runtime_types';
 import { ConnectorMappingsAttributesPartialRt } from './connector_mappings';
 
 describe('mappings', () => {
