+++ x-pack/plugins/cases/server/internal_attachments/index.ts
@@ -10,7 +10,7 @@ import { FileAttachmentMetadataRt } from '../../common/types/domain';
 import { LENS_ATTACHMENT_TYPE } from '../../common/constants/visualizations';
 import { FILE_ATTACHMENT_TYPE } from '../../common/constants';
 
-import { decodeWithExcessOrThrow } from '../../common/api';
+import { decodeWithExcessOrThrow } from '../common/runtime_types';
 import type { ExternalReferenceAttachmentTypeRegistry } from '../attachment_framework/external_reference_registry';
 import type { PersistableStateAttachmentTypeRegistry } from '../attachment_framework/persistable_state_registry';
