+++ x-pack/plugins/cases/server/routes/api/configure/post_configure.ts
@@ -6,7 +6,7 @@
  */
 
 import { ConfigurationRequestRt } from '../../../../common/types/api';
-import { decodeWithExcessOrThrow } from '../../../../common/api';
+import { decodeWithExcessOrThrow } from '../../../common/runtime_types';
 import { CASE_CONFIGURE_URL } from '../../../../common/constants';
 import { createCaseError } from '../../../common/error';
 import { createCasesRoute } from '../create_cases_route';
