+++ x-pack/plugins/cases/server/services/user_profiles/index.ts
@@ -16,7 +16,7 @@ import { DEFAULT_SPACE_ID } from '@kbn/spaces-plugin/common';
 import type { LicensingPluginStart } from '@kbn/licensing-plugin/server';
 import type { SuggestUserProfilesRequest } from '../../../common/types/api';
 import { SuggestUserProfilesRequestRt } from '../../../common/types/api';
-import { decodeWithExcessOrThrow } from '../../../common/api';
+import { decodeWithExcessOrThrow } from '../../common/runtime_types';
 import { Operations } from '../../authorization';
 import { createCaseError } from '../../common/error';
 import { LicensingService } from '../licensing';
